+++ src/app/page.tsx
@@ -1,8 +1,8 @@
 'use client'
 
-import shoppingRoutes from '@/flows/shopping/routes';
 import { useRouter } from 'next/navigation';
 import { useEffect } from 'react';
+import shoppingRoutes from '@/presentation/flows/shopping/routes';
 
 export default function HomePage() {
   const router = useRouter();
